requirements.txt
Flask
gunicorn
requests
openai
numpy
matplotlib
python-dotenv
markdown
langchain-openai
langchain-community
faiss-cpu
gspread
google-auth-oauthlib
pytz